Guilty Optimists: Kohya Refactor [Part A]  - Extract out Datasets to smaller files.
