+++ library/model_util.py
@@ -1133,48 +1133,3 @@ def load_vae(vae_id, dtype):
   return vae
 
 # endregion
-
-
-def make_bucket_resolutions(max_reso, min_size=256, max_size=1024, divisible=64):
-  max_width, max_height = max_reso
-  max_area = (max_width // divisible) * (max_height // divisible)
-
-  resos = set()
-
-  size = int(math.sqrt(max_area)) * divisible
-  resos.add((size, size))
-
-  size = min_size
-  while size <= max_size:
-    width = size
-    height = min(max_size, (max_area // (width // divisible)) * divisible)
-    resos.add((width, height))
-    resos.add((height, width))
-
-    # # make additional resos
-    # if width >= height and width - divisible >= min_size:
-    #   resos.add((width - divisible, height))
-    #   resos.add((height, width - divisible))
-    # if height >= width and height - divisible >= min_size:
-    #   resos.add((width, height - divisible))
-    #   resos.add((height - divisible, width))
-
-    size += divisible
-
-  resos = list(resos)
-  resos.sort()
-  return resos
-
-
-if __name__ == '__main__':
-  resos = make_bucket_resolutions((512, 768))
-  print(len(resos))
-  print(resos)
-  aspect_ratios = [w / h for w, h in resos]
-  print(aspect_ratios)
-
-  ars = set()
-  for ar in aspect_ratios:
-    if ar in ars:
-      print("error! duplicate ar:", ar)
-    ars.add(ar)
